src/requirements.txt
fastapi
uvicorn
python-dotenv
requests
beautifulsoup4
loguru
tiktoken
qdrant-client
